requirements.txt
requests
configparser
pyyaml
pathlib
zeep
argparse